# perf: drop the initial stack copy