This isn't much, but at least a good start (we did 3 overall copies, but seem like the initial one is unnecessary since we never touch that stack, so we do 2 of them now).

main:
```
$ touch t.py; time black -tpy310 --check t.py > /dev/null
would reformat t.py
Oh no! 💥 💔 💥
1 file would be reformatted.
black -tpy310 --check t.py > /dev/null  0,33s user 0,02s system 99% cpu 0,358 total
```
this branch:
```
$ touch t.py; time black -tpy310 --check t.py > /dev/null
would reformat t.py
Oh no! 💥 💔 💥
1 file would be reformatted.
black -tpy310 --check t.py > /dev/null  0,27s user 0,01s system 99% cpu 0,275 total
```
test script: https://gist.github.com/isidentical/2911e41262314e6f189149da45a10054