@ichard26 2021-12-05
```
parse-black/strings
===================

Mean +- std dev: [uhh-default] 67.0 ms +- 2.0 ms -> [uhh-initial] 227 ms +- 27 ms: 3.39x slower
Mean +- std dev: [uhh-default] 67.0 ms +- 2.0 ms -> [uhh-speedup] 154 ms +- 3 ms: 2.30x slower

parse-match-identifiers
=======================

Mean +- std dev: [uhh-default] 21.8 ms +- 2.7 ms -> [uhh-initial] 196 ms +- 23 ms: 9.01x slower
Mean +- std dev: [uhh-default] 21.8 ms +- 2.7 ms -> [uhh-speedup] 135 ms +- 10 ms: 6.19x slower

Geometric mean
==============

uhh-initial: 5.53x slower
uhh-speedup: 3.77x slower

❯ pyperf compare_to uhh-initial.json uhh-speedup.json -G
Faster (2):
- parse-black/strings: 227 ms +- 27 ms -> 154 ms +- 3 ms: 1.47x faster
- parse-match-identifiers: 196 ms +- 23 ms -> 135 ms +- 10 ms: 1.46x faster

Geometric mean: 1.47x faster
```